refactor: CardNumberField 비즈니스로직과 UI 분리
